Moved pycsvtosqliteroh to src directory
Move the package one lvl deeper to mayb prevent the error message:

No module named pycsvtosqliteroh.__main__; 'pycsvtosqliteroh' is a package and cannot be directly executed

This message seems to be related to the way how imports are handled in python.
But everything is the same as on my other modules, so is not fixed atm.

To use the module via cli run the following:

python -m pycsvtosqliteroh.main --file .\tests\testfiles\Printer.csv --database .\tests\Test.sqlite