Changed .gititnore file to keep one folder
The folder ./exportedfiles was not shown in remote repo.

Changed the folder exclusion on .gitignore to keep the directory.

This folder is used by the unittests to put the files in there, which
are  created from the sqlite table export, to than check if testfiles folder
and exportedfiles folder match.